requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
cffi==1.11.0
click==6.7
cryptography==2.0.3
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.5.3
pymongo==3.5.1
six==1.11.0
Werkzeug==0.12.2
